+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/viewmodel/AdvancedSettingsViewModel.kt
@@ -38,6 +38,17 @@ class AdvancedSettingsViewModel
                 }
             },
         ),
+        prioritizeSeekAccuracyState = State.PrioritizeSeekAccuracyState(
+            isChecked = settings.prioritizeSeekAccuracy.value,
+            onCheckedChange = {
+                settings.prioritizeSeekAccuracy.set(it, updateModifiedAt = false)
+                updatePrioritizeSeekAccuracyState()
+                analyticsTracker.track(
+                    AnalyticsEvent.SETTINGS_ADVANCED_PRIORITIZE_SEEK_ACCURACY,
+                    mapOf("enabled" to it),
+                )
+            },
+        ),
     )
 
     private fun onSyncOnMeteredCheckedChange(isChecked: Boolean) {
@@ -56,17 +67,31 @@ class AdvancedSettingsViewModel
         )
     }
 
+    private fun updatePrioritizeSeekAccuracyState() {
+        mutableState.value = mutableState.value.copy(
+            prioritizeSeekAccuracyState = mutableState.value.prioritizeSeekAccuracyState.copy(
+                isChecked = settings.prioritizeSeekAccuracy.value,
+            ),
+        )
+    }
+
     fun onShown() {
         analyticsTracker.track(AnalyticsEvent.SETTINGS_ADVANCED_SHOWN)
     }
 
     data class State(
         val backgroundSyncOnMeteredState: BackgroundSyncOnMeteredState,
+        val prioritizeSeekAccuracyState: PrioritizeSeekAccuracyState,
     ) {
 
         data class BackgroundSyncOnMeteredState(
             val isChecked: Boolean,
             val isEnabled: Boolean,
+            val onCheckedChange: (Boolean) -> Unit,
+        )
+
+        data class PrioritizeSeekAccuracyState(
+            val isChecked: Boolean,
             val onCheckedChange: (Boolean) -> Unit,
         )
     }